changed decimal helpers to long

+++ src/JK.Common.Tests/TypeHelpers/DecimalHelperTests.cs
@@ -8,7 +8,7 @@ namespace JK.Common.Tests.TypeHelpers
         [InlineData(0, 1.0)]
         [InlineData(1, 1.1)]
         [InlineData(25, 1.25)]
-        public void GetDecimalPart_Theories(int expected, decimal value)
+        public void GetDecimalPart_Theories(long expected, decimal value)
         {
             var actual = DecimalHelper.GetDecimalPart(value);
             Assert.Equal(expected, actual);
@@ -18,7 +18,7 @@ namespace JK.Common.Tests.TypeHelpers
         [InlineData(0, 0.12)]
         [InlineData(1, 1.54)]
         [InlineData(25, 25.75)]
-        public void GetWholePart_Theories(int expected, decimal value)
+        public void GetWholePart_Theories(long expected, decimal value)
         {
             var actual = DecimalHelper.GetWholePart(value);
             Assert.Equal(expected, actual);
